apply code fixes and ocd
