+++ packages/studiocms/src/runtime/index.ts
@@ -1,1 +1,3 @@
 export { createComponentProxy, dedent, transformHTML } from './AstroComponentProxy.js';
+
+export { decode } from './decoder/index.js'
