+++ packages/studiocms_devapps/package.json
@@ -61,7 +61,6 @@
     "astro-integration-kit": "catalog:",
     "cheerio": "^1.0.0",
     "turndown": "^7.2.0",
-    "html-entities": "^2.5.2",
     "sanitize-html": "^2.17.0"
   },
   "devDependencies": {
